move reducerUtils to @suite-utils
