+++ packages/suite/src/actions/suite/suiteActions.ts
@@ -1,5 +1,5 @@
 import TrezorConnect, { Device, DEVICE } from 'trezor-connect';
-import * as reducersUtils from '@suite-utils/reducers';
+import * as reducersUtils from '@suite-utils/reducerUtils';
 import * as deviceUtils from '@suite-utils/device';
 import { add as addNotification } from '@suite-actions/notificationActions';
 import * as routerActions from '@suite-actions/routerActions';
