+++ packages/suite/src/actions/wallet/transactionActions.ts
@@ -1,10 +1,10 @@
 import TrezorConnect, { AccountTransaction } from 'trezor-connect';
-import { getAccountTransactions } from '@suite/utils/wallet/reducerUtils';
+import { getAccountTransactions } from '@suite-utils/reducerUtils';
 import * as accountActions from '@wallet-actions/accountActions';
-import { SETTINGS } from '@suite/config/suite';
 import { TRANSACTION } from '@wallet-actions/constants';
 import { WalletAccountTransaction } from '@wallet-reducers/transactionReducer';
 import { db } from '@suite/storage';
+import { SETTINGS } from '@suite-config';
 import { Account } from '@wallet-types';
 import { Dispatch, GetState } from '@suite-types';
